[FARM-402] Add examples
