+++ lib/ui/playground/data/playground_atom_datasource.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 
 class PlayGroundAtomDataSource {
   List<Widget> getList() {
     return [
       //Add your atoms here
       Text('Atom widget 1'),
-      Text('Atom widget 2'),
+      Card(child: Text('Atom widget 2')),
       Text('Atom widget 4'),
     ];
   }
